ajax: User can delete his answer

+++ app/controllers/answers_controller.rb
@@ -11,11 +11,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author_of?(answer)
-      answer.destroy
-      flash[:notice] = 'The answer is successfully deleted.'
-    end
-    redirect_to answer.question
+    answer.destroy if current_user.author_of?(answer)      
   end
 
   def update
